A hack to loosen stream check strictness... discarding the first header on resync. That at least enables playback of the test file of bug 1889051 ... but may introduce other false positives again. Actually, the problem of 1889051 is a false positive for the fist header; which _passes_ the text for a following valid header after that frame. What a coincidence... Should I test for the next 2 frames? MPlayer doesn't show the short distortion that mpg123 shows in decoding the first (invalid, but strangely decodable) frame.
Guess the best to do is to finally rewrite that frame reading/searching function to be more consistent and tunable for
	- bytes of junk to skip
	- resync limit (we have that already)
	- number of frames/headers readahead for assuring a valid stream

I would like to have a 1.2.1 release with just the last bugfix (endless loop in resync) and perhaps a variant of the current hack after I spent some more thought and testing on it (does mpg123 still pass the nasty test files that made me make the resync that strickt?).
Then, for 1.3.0, the tunable resync should enter the scene.



git-svn-id: svn://svn.orgis.org/mpg123/trunk/src/libmpg123@1375 35dc7657-300d-0410-a2e5-dc2837fedb53