Do not anticipate track end on reaching an expected number of frames (from Xing header). Only end when there is nothing to read anymore.
This is to support http streams that like to prepend an intro piece with Xing header (that tells mpg123 that the stream is about to end in a few seconds).
An optimal solution would check for that case in the parser (seeing that we are done with a sub-track and thus look out anew for a first frame and Xing/Lame header, tags), but there is potential trouble with gapless state... and audio_start .. seek index. Well, one might get away with non-seekable streams, but when you dump the webradio, you got a seekable file on disk.



git-svn-id: svn://svn.orgis.org/mpg123/trunk/src/libmpg123@2427 35dc7657-300d-0410-a2e5-dc2837fedb53